Add homepage social sharing metadata
